# doc/openapi.yaml
openapi: 3.1.0
info:
  title: tlstap API
  version: 0.3.0
  description: |
    REST API exposed by the tlstap TLS intercepting proxy when the `dbdump`
    and/or `tamper` interceptors are active and `"api": {"listen": "..."}` is
    set in `config.json`.

    **Base path:** All endpoints are reachable at the canonical path
    `/api/i/<interceptor-name>/<endpoint>` as well as the proxy-scoped path
    `/<proxyName>/api/i/<interceptor-name>/<endpoint>`.

    **Errors:** All error responses use `Content-Type: application/json` and the
    body `{"error": "<message>"}` with an appropriate HTTP status code.

    **Request validation:** a `400` on a JSON endpoint reliably means the body failed
    to *parse* (malformed JSON, wrong field type). A `required` field in a request
    schema below documents the field a well-formed request is expected to supply —
    omitting it is not otherwise rejected; the handler just runs with that field's zero
    value (`0`, `""`, `false`, or an empty array), which is occasionally the correct,
    documented behavior (e.g. `/chunk`'s `chunks`) and occasionally just an
    unintentionally-permissive gap. Don't rely on a missing required field producing
    an error.

    **dbdump vs. tamper:** `dbdump` passively records traffic to SQLite and
    exposes a pull/replay history API. `tamper` is a separate, live-only
    interceptor that lets a connected control client pause, edit, drop, or
...
    `tamper` state can be correlated by `conn` alone.
servers:
...
paths:
...
  /api/i/tamper/control:
    get:
      tags: [tamper]
      summary: Single control connection for live traffic interception
      description: |
        Upgrades the connection to WebSocket (`Upgrade: websocket`). Only one
        control connection is allowed at a time; a second attempt is rejected
        with an `error` frame and the connection is closed. All commands and
        notifications for every stream flow through this one connection —
        per-stream mode toggles and held-chunk resolution included, not just
        lifecycle events.

        With no control connection active, every stream on this interceptor
        behaves as pure pass-through (watch mode, no watchers, zero overhead)
        — unless the interceptor is configured with `hold-until-connected`,
        in which case every chunk is held (bounded by `hold-timeout-ms`)
        until a control client connects, so traffic is never silently missed
        while nothing is attached yet.

        A held chunk is appended to a growing, editable byte buffer kept per
        `(conn, direction)` — not addressed by its own id. No bytes ever flow
        over this connection — `held` and `stream-list` are metadata only. To
        read a direction's currently-held bytes, connect to that stream's
        `/api/i/tamper/watch` connection and send a `peek` command.

        **Proxy → client** (JSON text frame):
        - ```json
          {"type": "stream-created", "conn": N, "src": "...", "dst": "..."}
          ```
        - ```json
          {"type": "stream-terminated", "conn": N}
          ```
        - ```json
          {"type": "held", "conn": N, "direction": N, "offset": N, "length": N, "time": N}
          ```
          Sent when a stream in intercept mode appends a new chunk to a
          direction's held buffer. `direction`: `0` = client→server, `1` =
          server→client. `offset`/`length` describe just this new chunk
          (where it starts within the buffer, and its size) — not the
          buffer's total size — so a client can extend its own
          locally-tracked view of the buffer incrementally rather than
          re-`peek`ing on every arrival; `offset` also doubles as a
          consistency check against that local view. `time` is milliseconds
          since the Unix epoch. Fetch the actual bytes via `peek` on
          `/watch`.
        - ```json
          {"type": "stream-list", "streams": [{"conn": N, "src": "...", "dst": "...", "intercepting": bool, "pending": [{"direction": N, "chunks": N, "length": N}]}]}
          ```
          Reply to a `list-streams` command. `pending` has 0-2 entries — one
          per direction with something currently held, summarized as chunk
          count + total byte length (not itemized per chunk) — this is what
          makes reconnecting to control at any time a full resync rather than
          just a feed of future events, e.g. after `hold-until-connected`
          held traffic before anyone was connected to see the original
          `held` push.
        - ```json
          {"type": "ok", "command": "set-auto-intercept" | "set-mode" | "release" | "drop-connection"}
          ```
          Acknowledges that command succeeded. Each of those four commands
          gets exactly one `ok` or `error` reply — never both, never neither.
          `script-log` (below) is the one command type that gets neither on
          success — it's fire-and-forget.
        - ```json
          {"type": "error", "message": "..."}
          ```
          Sent for: an unrecognized command type; missing required fields;
          `set-mode`/`release`/`drop-connection` against an unknown stream;
          `release` whose `prefix_length` exceeds the buffer's actual current
          length (e.g. a hold-timeout already flushed it) or whose
          `release_chunks` exceeds the resulting chunk count; or `script-log`
          with a `level` other than `"log"`/`"error"`. The connection is not
          closed.

        **Client → proxy** (JSON text frame; `release` with `"edited": true`
        must be immediately followed by a binary frame with the replacement
        bytes for `prefix_length`):
        - ```json
          {"type": "set-auto-intercept", "enabled": bool}
          ```
          Sets the mode newly-established connections start in. Does not
          affect streams that already exist.
        - ```json
          {"type": "set-mode", "conn": N, "intercepting": bool}
          ```
          Switches one stream between watch and intercept mode. Switching to
          `false` immediately force-releases (forwards, unmodified)
          everything currently held for that stream, rather than leaving it
          to time out.
        - ```json
          {"type": "release", "conn": N, "direction": N, "edited": bool, "prefix_length": N, "bounds": [N, ...], "release_chunks": N, "action": "forward" | "drop"}
          ```
          Combines an optional edit with an optional release of one
          direction's held buffer, in one atomic step. `edited: false` means
          "release `release_chunks` chunks of the buffer exactly as it
          stands" (`prefix_length`/`bounds` are ignored). `edited: true`
          replaces the buffer's first `prefix_length` bytes — as the client
          last knew them — with the bytes in the following binary frame,
          whose internal chunk boundaries are given by `bounds` (a list of
          byte offsets, `bounds[0]` always `0`), before releasing.
          `prefix_length` anchors the edit against the buffer's actual
          current length: a stale value (e.g. a hold-timeout flushed the
          buffer in the meantime) is rejected rather than silently
          misapplied. `release_chunks` `0` means "edit only, release
          nothing" (a real supported mode, not just a corner case).
          `action: "forward"` releases the real bytes; `action: "drop"`
          removes the released chunks from the buffer without forwarding
          them. Unresolved held buffers are also auto-released, unmodified,
          after the configured `hold-timeout-ms` (or immediately, if this
          control connection disconnects).
        - ```json
          {"type": "drop-connection", "conn": N, "direction": N}
          ```
          Terminates the connection outright, independent of `release` —
          dropping the connection isn't parameterized by buffer content the
          way an edit/release is, so it's its own command rather than a
          third `release` action value. `direction` just identifies which
          buffer the client was looking at; either direction's buffer kills
          the whole connection.
        - ```json
          {"type": "list-streams"}
          ```
          Requests a `stream-list` snapshot of every currently active stream.
        - ```json
          {"type": "script-log", "level": "log" | "error", "text": "..."}
          ```
          Pushes one already-formatted line from a running script's log (see
          the "Scripts" sub-tab / `scriptRuntime.js`) for optional
          server-side persistence — see `GET /api/i/tamper/log-file`. A
          no-op if the interceptor wasn't configured with `log-file`.
          `text` may itself span multiple lines. Fire-and-forget: no `ok`
          reply on success, only `error` for an invalid `level`.
      operationId: tamperControl
      responses:
        '101':
          description: Switching Protocols — WebSocket connection established
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/i/tamper/watch:
    get:
      tags: [tamper]
      summary: Live mirror of one stream's traffic, plus on-demand reads of held buffers
      description: |
        Upgrades the connection to WebSocket (`Upgrade: websocket`). Any
        number of clients may attach to the same stream at once. This
        connection is never a decision-maker — forward/edit/drop/
        drop-connection decisions only ever happen on
        `/api/i/tamper/control`; this connection only ever reads.

        **Live mirror** (server → client, unsolicited, best-effort — a client
        that can't keep up simply misses frames rather than slowing down the
        proxy): every chunk actually forwarded on the stream (both watch-mode
        pass-through and released held buffers), as two frames:
        1. Text frame (JSON metadata):
           ```json
           {"direction": N, "time": N, "length": N}
           ```
        2. Binary frame: raw chunk bytes

        **`peek`** (client → server): the only way to read a direction's
        currently-held (not-yet-released) bytes. Fetches the *whole* buffer
        for that direction, optionally sliced.
        ```json
        {"type": "peek", "direction": N, "offset": N, "length": N}
        ```
        - `direction` — required, `0` or `1`.
        - `offset`, `length` — optional byte-range slice of the buffer
          (`length` `0` or omitted means "to the end").

        Reply is exactly one of the following pairs, terminated by a
        `peek-done` frame:
        1. Text frame (JSON metadata):
           ```json
           {"type": "pending", "direction": N, "time": N, "offset": N, "length": N, "total_length": N, "bounds": [N, ...]}
           ```
           `offset`/`length` describe the slice actually returned;
           `total_length` and `bounds` always describe the *whole* buffer
           regardless of what was sliced, since a resync always wants the
           complete picture and both are cheap. A direction with nothing
           currently held is a normal zero-length reply here, not an error.
        2. Binary frame: the (possibly sliced) buffer bytes

        Then:
        ```json
        {"type": "peek-done"}
        ```

        If `direction` was invalid, the reply is a single error frame
        instead:
        ```json
        {"type": "error", "message": "..."}
        ```
      operationId: tamperWatch
      parameters:
        - name: conn
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ConnID of the stream to mirror
      responses:
        '101':
          description: Switching Protocols — WebSocket connection established
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No stream with the given conn id is currently active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
  responses:
    BadRequest:
      description: Invalid or malformed request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...